# docs/swagger.yml
openapi: 3.0.0
info:
  title: User Registration and Login API
  description: API for user registration (temporary and permanent) and login
  version: 1.0.0
servers:
  - url: http://localhost/api
    description: Local development server
paths:
...
  /auth/verify/{token}:
    get:
      summary: Permanent user registration
      description: Verify user email and complete registration
      tags:
        - Authentication
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Verification token sent in email
      responses:
        '200':
          description: User registration completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registration completed successfully."
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid verification token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid verification token"
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Test User"
        email:
          type: string
          format: email
          example: "test@example.com"
        email_verified_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-05-04T03:30:45.000000Z"
        created_at:
          type: string
          format: date-time
          example: "2025-05-04T03:25:45.000000Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-05-04T03:30:45.000000Z"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT